Atualizando Knexfile e Migration de Usuario
