+++ backend/src/index.js
@@ -6,7 +6,9 @@ const routes = require('./routes');
 const app = express();
 
 app.use(cors());
+
 app.use(express.json());
+
 app.use(routes);
 
 app.listen(3333);
