--- src/routes.js
@@ -1,11 +1,0 @@
-const express = require('express');
-
-
-const routes = express.Router();
-
-routes.get('/',(request,response)=>{
-    return response.json({message: "Miauna Papai"})
-});
-
-
-module.exports = routes;
